prevent deletion of .htaccess issue #91

+++ data/inc/deletefile.php
@@ -17,22 +17,27 @@ defined('IN_PLUCK') or exit('Access denied!');
 
 //Check if image exists.
 if (file_exists('files/'.$var1)) {
-	//First check if there isn't an item with the same name in the trashcan.
-	if (!file_exists('data/trash/files/'.$var1)) {
-		//Move the image to the trashcan.
-		copy('files/'.$var1, 'data/trash/files/'.$var1);
-		chmod('data/trash/files/'.$var1, 0777);
-		unlink('files/'.$var1);
 
-		//Redirect user.
-		show_error($lang['trashcan']['moving_item'], 3);
-		redirect('?action=files', 0);
-	}
+	if ($var1 == ".htaccess"){
+		show_error($lang['trashcan']['notallowed'], 1);
+	} else {
+		//First check if there isn't an item with the same name in the trashcan.
+		if (!file_exists('data/trash/files/'.$var1)) {
+			//Move the image to the trashcan.
+			copy('files/'.$var1, 'data/trash/files/'.$var1);
+			chmod('data/trash/files/'.$var1, 0777);
+			unlink('files/'.$var1);
 
-	//If there is an item with the same name in the trashcan: display error.
-	else {
-		show_error($lang['trashcan']['same_name'], 1);
-		redirect('?action=files', 3);
+			//Redirect user.
+			show_error($lang['trashcan']['moving_item'], 3);
+			redirect('?action=files', 0);
+		}
+
+		//If there is an item with the same name in the trashcan: display error.
+		else {
+			show_error($lang['trashcan']['same_name'], 1);
+			redirect('?action=files', 3);
+		}
 	}
 }
 ?>
